Added color picker
