+++ Podfile
@@ -4,7 +4,7 @@
 target 'Booktrix' do
   # Comment the next line if you're not using Swift and don't want to use dynamic frameworks
   use_frameworks!
-
+  pod 'IGColorPicker'
   # Pods for Booktrix
 
   target 'BooktrixTests' do
